Use MUnit in ResultResponseSuite
